feat: more boilerplate
Note: boilerplate needs some TLC to change it from a triangle to a
circle

Also, having a markdown file have the extension '.markdown' is for plebs. Maybe this is the result of using a programming language that is far older than the dinosaurs themselves to boilerplate the project, but '.md' really should be the way to go.